Update information
(1) Identifcation merging
If we have different Ids from diferent runs, and they localized on the same peak time window. In this case, we need to merge the whole ids.

(2) Duplicated quantified Peak
If we have the same quantified peak from FlashLFQ and IsoTracker, then we will check is the ids, making sure the id from FlashLFQ are included in those of IsoTracke. Finally we eliminate the Peak from FlashLFQ

(3)Tester for item 1